:zap: review 신고하기 기능 구현

+++ src/main/java/com/farmer/cornfarmer/src/review/model/PostReportRes.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class PostReportRes {
-    private int reportIdx;
+    private long reportIdx;
 }
